factor helper namespace into separate header file (#7)
