+++ base/threading/simple_thread.h
@@ -2,6 +2,8 @@
 #define BASE_THREADING_SIMPLE_THREAD_H_
 
 #include <functional>
+#include <tuple>
+#include <utility>
 
 #include "base/threading/thread.h"
